requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
sentence-transformers==2.2.2
faiss-cpu==1.7.3
PyPDF2==3.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0
